requirements.txt
kafka-python==2.2.2
pandas==2.2.0
pyspark==3.5.0
